vinyl-label-printer/requirements.txt
PyQt6>=6.4.0
reportlab>=4.0.0
openpyxl>=3.1.0
pdf2image>=1.17.0
Pillow>=10.0.0
# System dependency (not pip-installable):
#   Linux:   sudo apt install poppler-utils
#   macOS:   brew install poppler
#   Windows: download DLLs from https://github.com/oschwartz10612/poppler-windows
#            and add the bin/ folder to PATH
